Lots of changes to tx module
Changed all msb's for all bit fields to end in "H" so that if the number of bits changes, the calculations in the PASM area will still be correct.
Insted of using a single jump table for all 16 combinations of II and QQ, there are now 3 jump tables, one for input mode, and two for output mode.
This makes things easier to read and reduces copy-paste code.
Added lots of comments. This compiles but still hasn't been tested yet.